requirements.txt
alembic==0.7.6
aniso8601==1.0.0
Flask==0.10.1
Flask-Migrate==1.4.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
SQLAlchemy==1.0.6
Werkzeug==0.10.4
wheel==0.24.0
